requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
noisereduce>=3.0.0
scipy>=1.10.0
customtkinter>=5.2.0
pystray>=0.19.0
Pillow>=10.0.0
